requirements.txt
requests==2.20.0
Pillow==5.3.0
